requirements.txt
fastapi>=0.68.1
uvicorn>=0.15.0
python-multipart>=0.0.5
numpy>=1.21.2
opencv-python>=4.5.5.64
deepface>=0.0.75
Pillow>=8.3.2 